pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "easy_whitelist"
authors = [
  {name = "qiqilelebaobao"},
  {name = "qiqilelebaobao", email = "qiqilelebaobao@163.com"}
]

readme = "README.md"
license = {text = "Apache License 2.0"}
classifiers = [
  "License :: OSI Approved :: Apache Software License", 
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.6",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14"
  ]
dynamic = ["version", "description"]
dependencies = [
  "tencentcloud-sdk-python",
]
maintainers = [
  {name = "qiqilelebaobao", email = "qiqilelebaobao@163.com"}
]

keywords = ["automation", "whitelist", "acl", "security-groups", "alibaba-cloud", "tencent-cloud", "security-tools", "security-group-rule"]

[project.urls]
Homepage = "https://github.com/qiqilelebaobao/easy_whitelist"

[project.scripts]
easy = "easy_whitelist.__main__:main"